Split SpecStyleGenerator into build, shapes and track compilation
Twenty-two private functions in one class, and every one of them opened with the
same three arguments: the builder, the config and the layout. That is what a
class is for, so there are now three.

SpecTracks compiles animation tracks into Lottie properties. SpecShapes builds
the rect, ellipse, polygon and path elements. SpecBuild holds one render's
builder, config and layout, owns the other two, and draws the logo, image and
text elements itself. SpecStyleGenerator keeps only its public shape: generate(),
lastWarnings and fromResource().

Two signatures stayed long after losing those three arguments, and both were
carrying a group that always travels together: buildShape's size, corner,
growFrom and repeat are now ShapeAttrs, and buildTextMask's rest, width, size and
justify are PlacedText.

Verified by the SpecPort suites, which compare each spec style byte-for-byte
against the compiled style it replaced, and by dumping all twelve compiled styles
across 36 configurations -- unchanged.
